+++ ucloud/resource_ucloud_lb_attachment.go
@@ -8,6 +8,7 @@ import (
 	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/hashicorp/terraform/helper/validation"
 
+	"github.com/ucloud/ucloud-sdk-go/services/ulb"
 	"github.com/ucloud/ucloud-sdk-go/ucloud"
 )
 
@@ -137,17 +138,31 @@ func resourceUCloudLBAttachmentUpdate(d *schema.ResourceData, meta interface{}) 
 
 func resourceUCloudLBAttachmentRead(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(*UCloudClient)
-
-	lbId := d.Get("load_balancer_id").(string)
-	listenerId := d.Get("listener_id").(string)
-
-	backendSet, err := client.describeBackendById(lbId, listenerId, d.Id())
-	if err != nil {
-		if isNotFoundError(err) {
-			d.SetId("")
-			return nil
+	var err error
+	var lbId string
+	var listenerId string
+	var backendSet *ulb.ULBBackendSet
+	if v, ok := d.GetOk("load_balancer_id"); ok {
+		listenerId = d.Get("listener_id").(string)
+		backendSet, err = client.describeBackendById(v.(string), listenerId, d.Id())
+		if err != nil {
+			if isNotFoundError(err) {
+				d.SetId("")
+				return nil
+			}
+			return fmt.Errorf("error on reading lb attachment %q, %s", d.Id(), err)
 		}
-		return fmt.Errorf("error on reading lb attachment %q, %s", d.Id(), err)
+
+		d.Set("load_balancer_id", v)
+		d.Set("listener_id", listenerId)
+	} else {
+		backendSet, lbId, listenerId, err = client.describeBackendByOneId(d.Id())
+		if err != nil {
+			return fmt.Errorf("error on parsing lb attachment %q, %s", d.Id(), err)
+		}
+
+		d.Set("load_balancer_id", lbId)
+		d.Set("listener_id", listenerId)
 	}
 
 	d.Set("resource_id", backendSet.ResourceId)